requirements.txt
streamlit
tensorflow
keras
opencv-python
Pillow
pandas
numpy
plotly
face_recognition
scikit-learn